move idam client and proxy into their own folder
